Added junction categories

+++ api_service/routes/item.js
@@ -15,17 +15,17 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.itemRouter = void 0;
 const express_1 = __importDefault(require("express"));
 exports.itemRouter = express_1.default.Router();
-const { sequelize, Item, Manufacturer } = require("../models");
+const { sequelize, Item, Manufacturer, Category, ItemCategory } = require("../models");
 exports.itemRouter.use(express_1.default.json());
 exports.itemRouter.use(express_1.default.urlencoded({ extended: true }));
 // Read
 exports.itemRouter.get('/', (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const items = yield Item.findAll({
-            include: {
-                model: Manufacturer,
-                as: 'manufacturer'
-            }
+            include: [
+                { model: Manufacturer, as: 'manufacturer' },
+                { model: Category, through: ItemCategory, as: 'categories' }
+            ]
         });
         return res.json(items);
     }
@@ -39,7 +39,10 @@ exports.itemRouter.get('/:id', (req, res) => __awaiter(void 0, void 0, void 0, f
     try {
         const item = yield Item.findOne({
             where: { id: req.params.id },
-            include: { model: Manufacturer, as: 'manufacturer' }
+            include: [
+                { model: Manufacturer, as: 'manufacturer' },
+                { model: Category, through: ItemCategory, as: 'categories' }
+            ]
         });
         return res.json(item);
     }
@@ -52,6 +55,14 @@ exports.itemRouter.get('/:id', (req, res) => __awaiter(void 0, void 0, void 0, f
 exports.itemRouter.post('/', (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const newItem = yield Item.create(req.body);
+        if (req.body.categories) {
+            for (let categoryId of req.body.categories) {
+                const category = yield Category.findByPk(categoryId);
+                if (category) {
+                    yield newItem.addCategory(category);
+                }
+            }
+        }
         return res.json(newItem);
     }
     catch (err) {
@@ -65,8 +76,23 @@ exports.itemRouter.put('/:id', (req, res) => __awaiter(void 0, void 0, void 0, f
         yield Item.update(req.body, { where: { id: req.params.id } });
         const updatedItem = yield Item.findOne({
             where: { id: req.params.id },
-            include: { model: Manufacturer, as: 'manufacturer' }
+            include: [
+                { model: Manufacturer, as: 'manufacturer' },
+                { model: Category, through: ItemCategory, as: 'categories' }
+            ]
         });
+        if (req.body.categories) {
+            const currentCategories = yield updatedItem.getCategories();
+            for (let category of currentCategories) {
+                yield updatedItem.removeCategory(category);
+            }
+            for (let categoryId of req.body.categories) {
+                const category = yield Category.findByPk(categoryId);
+                if (category) {
+                    yield updatedItem.addCategory(category);
+                }
+            }
+        }
         return res.json(updatedItem);
     }
     catch (err) {
@@ -77,7 +103,14 @@ exports.itemRouter.put('/:id', (req, res) => __awaiter(void 0, void 0, void 0, f
 // Delete item
 exports.itemRouter.delete('/:id', (req, res) => __awaiter(void 0, void 0, void 0, function* () {
     try {
-        yield Item.destroy({ where: { id: req.params.id } });
+        const item = yield Item.findByPk(req.params.id);
+        if (item) {
+            const categories = yield item.getCategories();
+            for (let category of categories) {
+                yield item.removeCategory(category);
+            }
+            yield Item.destroy({ where: { id: req.params.id } });
+        }
         return res.json({ message: `Item with id ${req.params.id} deleted.` });
     }
     catch (err) {
